+++ lib/std/posix/io_uring.nim
@@ -457,14 +457,14 @@ proc `+`(p: pointer; i: uint32): pointer =
 proc uringMap(offset: Off; fd: FileHandle; begin: uint32;
                count: uint32; typSize: int): pointer {.raises, tags: [].} =
   let size = int(begin + count * typSize.uint32)
-  result = mmap(nil, size, PROT_READ or PROT_WRITE, MAP_SHARED or MAP_POPULATE,
+  result = mmap(nil, csize_t(size), PROT_READ or PROT_WRITE, MAP_SHARED or MAP_POPULATE,
                 fd.cint, offset)
   if mmapFailed(result):
     raiseOSError(osLastError(), "io_uring uringMap failed")
 
 proc uringUnmap(p: pointer; size: int) {.raises, tags: [].} =
   ## interface to tear down some memory (probably mmap'd)
-  let code = munmap(p, size)
+  let code = munmap(p, csize_t(size))
   if code < 0:
     raiseOSError(osLastError(), "io_uring uringUnmap failed")
